+++ gradio/components/browser_state.py
@@ -61,22 +61,9 @@ class BrowserState(Component):
         Parameters:
             value: Value to store in local storage
         Returns:
-            Value converted to JSON-serializable format
+            Passes value through unchanged
         """
-        return self.make_serializable(value)
-
-    def make_serializable(self, obj: Any) -> dict | list | str:
-        if isinstance(obj, list):
-            return [self.make_serializable(item) for item in obj]
-        elif isinstance(obj, dict):
-            return {k: self.make_serializable(v) for k, v in obj.items()}
-        elif hasattr(obj, "get_config"):
-            try:
-                if hasattr(obj, "value"):
-                    return obj.value
-            except Exception:
-                return str(obj)
-        return obj
+        return value
 
     def api_info(self) -> dict[str, Any]:
         return {"type": {}, "description": "any json-serializable value"}
